Custom hook for focus trap & other fixes
